requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiosqlite==0.19.0
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1

# OpenVINO Whisper (using GenAI - simpler API, no PyTorch conflicts)
openvino-genai>=2024.3.0
librosa>=0.10.0
soundfile>=0.12.0
huggingface-hub>=0.20.0
